+++ src/Interfaces/order/OrderLine.ts
@@ -1,5 +1,6 @@
 import ProductVariant from '../product/ProductVariant';
 import Order from './Order';
+import Stock from '../stock/Stock';
 
 export default interface OrderLine {
     id?: string;
@@ -23,4 +24,5 @@ export default interface OrderLine {
     // belongs to
     order?: Order | null;
     productVariant?: ProductVariant | null;
+    stock?: Array<Stock>;
 }