+++ src/Interfaces/subscription/SubscriptionAddon.ts
@@ -1,0 +1,11 @@
+import SubscriptionAddonPricingPlan from './SubscriptionAddonPricingPlan';
+
+export default interface SubscriptionAddon {
+    id?: string;
+    createdAt?: Date | string;
+    updatedAt?: Date | string;
+    archivedAt?: Date | string;
+    name?: string;
+    stripeId?: string;
+    subscriptionAddonPricingPlans?: Array<SubscriptionAddonPricingPlan>;
+}