+++ src/Models/basket/Transformers/BasketTransformer.ts
@@ -1,6 +1,6 @@
 import Basket from '../Basket';
 import BasketLine from '../BasketLine';
-import { APIBasket as APIBasketInterface } from '../../../Interfaces/basket/Basket';
+import { Basket as BasketInterface } from '../../../Interfaces/basket/Basket';
 import BasketLineTransformer from './BasketLineTransformer';
 import BaseTransformer from '../../../BaseTransformer';
 
@@ -10,16 +10,16 @@ export default class BasketTransformer extends BaseTransformer {
      *
      * @param basket
      */
-    mapData(basket: APIBasketInterface): Basket {
+    mapData(basket: BasketInterface): Basket {
         return new Basket({
             id: basket.id,
-            tokensTotal: basket.tokens_total,
-            createdAt: basket.created_at,
-            updatedAt: basket.updated_at,
+            totalTokens: basket.totalTokens,
+            createdAt: basket.createdAt,
+            updatedAt: basket.updatedAt,
             basketLines: this.includeBasketLines(basket),
         });
     }
-    includeBasketLines(basket: APIBasketInterface): Array<BasketLine> {
+    includeBasketLines(basket: BasketInterface): Array<BasketLine> {
         return this.collection(basket, 'basket_lines', new BasketLineTransformer());
     }
 }