+++ src/Models/product/ProductListLine.ts
@@ -1,0 +1,7 @@
+import ProductListLineInterface from '../../Interfaces/product/ProductListLine';
+
+export default class ProductListLine {
+    constructor(parameters: ProductListLineInterface) {
+        Object.assign(this, parameters);
+    }
+}