+++ src/Models/reasonForEnd/ReasonForEnd.ts
@@ -1,0 +1,7 @@
+import ReasonForEndInterface from '../../Interfaces/reasonForEnd/ReasonForEnd';
+
+export default class ReasonForEnd {
+    constructor(parameters: ReasonForEndInterface) {
+        Object.assign(this, parameters);
+    }
+}