+++ src/Models/role/Role.ts
@@ -1,0 +1,7 @@
+import RoleInterface from '../../Interfaces/role/Role';
+
+export default class Role {
+    constructor(parameters: RoleInterface) {
+        Object.assign(this, parameters);
+    }
+}