+++ src/Models/user/UserSubscriptionAddon.ts
@@ -1,0 +1,7 @@
+import UserSubscriptionAddonInterface from '../../Interfaces/user/UserSubscriptionAddon';
+
+export default class UserSubscriptionAddon {
+    constructor(parameters: UserSubscriptionAddonInterface) {
+        Object.assign(this, parameters);
+    }
+}