+++ src/Resources/Guests/index.ts
@@ -3,12 +3,14 @@ import Baskets from './Baskets';
 import Gifts from './Gifts';
 import ShippingAddons from './ShippingAddons';
 import ReasonsForEnd from './ReasonsForEnd';
+import Subscriptions from './Subscriptions';
 
 export interface GuestsResources {
     baskets: Baskets;
     gifts: Gifts;
     shippingAddons: ShippingAddons;
     reasonsForEnd: ReasonsForEnd;
+    subscriptions: Subscriptions;
 }
 
 export function loadGuestsResources(client: Client): GuestsResources {
@@ -17,5 +19,6 @@ export function loadGuestsResources(client: Client): GuestsResources {
         gifts: new Gifts(client),
         shippingAddons: new ShippingAddons(client),
         reasonsForEnd: new ReasonsForEnd(client),
+        subscriptions: new Subscriptions(client),
     };
 }