+++ src/Resources/Wacc/Roles/Roles.ts
@@ -1,0 +1,14 @@
+import AbstractResource from '../../AbstractResource';
+import { HttpClientConfig, HttpClientResponse } from '../../../Interfaces/HttpClient';
+
+export default class Roles extends AbstractResource {
+    initialise(): void {
+        this.resourcePath = '/roles';
+        this.defaultAccess = 'wacc';
+    }
+
+    public all: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
+        method: 'GET',
+        path: '/',
+    });
+}